formatted

+++ src/DataTypes.sol
@@ -5,7 +5,7 @@ contract DataTypes {
     bool public hi;
     bool public hello = true;
 
-    int public num = -7;
+    int256 public num = -7;
     uint256 public num1 = 55;
 
     string public String = "string";
